[CHORE]#18 : data-remote 모듈 구현 및 network 모듈 migration
