--- core/network/src/main/kotlin/com/bff/wespot/network/extensions/DateTimeExtensions.kt
@@ -1,7 +1,0 @@
-package com.bff.wespot.network.extensions
-
-import java.time.LocalDateTime
-import java.time.format.DateTimeFormatter
-
-internal fun String.toISOLocalDateTime(): LocalDateTime =
-    LocalDateTime.parse(this, DateTimeFormatter.ISO_DATE_TIME)
