--- core/network/src/main/kotlin/com/bff/wespot/network/model/auth/request/ConsentsDto.kt
@@ -1,8 +1,0 @@
-package com.bff.wespot.network.model.auth.request
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class ConsentsDto (
-    val marketing: Boolean,
-)
