--- core/network/src/main/kotlin/com/bff/wespot/network/model/auth/request/KakaoAuthTokenDto.kt
@@ -1,9 +1,0 @@
-package com.bff.wespot.network.model.auth.request
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class KakaoAuthTokenDto(
-    val socialType: String,
-    val identityToken: String,
-)
