--- core/network/src/main/kotlin/com/bff/wespot/network/model/auth/request/SignUpDto.kt
@@ -1,14 +1,0 @@
-package com.bff.wespot.network.model.auth.request
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class SignUpDto(
-    val name: String,
-    val schoolId: Int,
-    val grade: Int,
-    val classNumber: Int,
-    val gender: String,
-    val signUpToken: String,
-    val consents: ConsentsDto,
-)
