--- core/network/src/main/kotlin/com/bff/wespot/network/model/auth/response/SchoolListDto.kt
@@ -1,8 +1,0 @@
-package com.bff.wespot.network.model.auth.response
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class SchoolListDto(
-    val schools: List<SchoolDto>
-)
