--- core/network/src/main/kotlin/com/bff/wespot/network/model/message/request/MessageTypeDto.kt
@@ -1,8 +1,0 @@
-package com.bff.wespot.network.model.message.request
-
-enum class MessageTypeDto {
-    RECEIVED,
-    SENT,
-}
-
-fun MessageTypeDto.name() = this.name.lowercase()
