--- core/network/src/main/kotlin/com/bff/wespot/network/model/message/request/SentMessageDto.kt
@@ -1,13 +1,0 @@
-package com.bff.wespot.network.model.message.request
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class SentMessageDto (
-    val receivedId: Int,
-    val schoolId: Int,
-    val grade: Int,
-    val group: Int,
-    val content: String,
-    val sender: String,
-)
