--- core/network/src/main/kotlin/com/bff/wespot/network/model/message/response/MessageIdDto.kt
@@ -1,8 +1,0 @@
-package com.bff.wespot.network.model.message.response
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class MessageIdDto (
-    val id: String,
-)
