--- core/network/src/main/kotlin/com/bff/wespot/network/model/message/response/MessageListDto.kt
@@ -1,15 +1,0 @@
-package com.bff.wespot.network.model.message.response
-
-import com.bff.wespot.model.message.response.MessageList
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class MessageListDto (
-    val messages: List<MessageDto>,
-    val isLastPage: Boolean,
-) {
-    fun toMessageList(): MessageList = MessageList(
-        messages = messages.map { it.toMessage() },
-        isLastPage = isLastPage,
-    )
-}
