+++ data-remote/src/main/kotlin/com/bff/wespot/data/remote/extensions/DateTimeExtensions.kt
@@ -1,0 +1,7 @@
+package com.bff.wespot.data.remote.extensions
+
+import java.time.LocalDateTime
+import java.time.format.DateTimeFormatter
+
+fun String.toISOLocalDateTime(): LocalDateTime =
+    LocalDateTime.parse(this, DateTimeFormatter.ISO_DATE_TIME)
