+++ data-remote/src/main/kotlin/com/bff/wespot/data/remote/model/auth/request/ConsentsDto.kt
@@ -1,0 +1,8 @@
+package com.bff.wespot.data.remote.model.auth.request
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class ConsentsDto (
+    val marketing: Boolean,
+)
