+++ data-remote/src/main/kotlin/com/bff/wespot/data/remote/model/auth/request/KakaoAuthTokenDto.kt
@@ -1,0 +1,9 @@
+package com.bff.wespot.data.remote.model.auth.request
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class KakaoAuthTokenDto(
+    val socialType: String,
+    val identityToken: String,
+)
