+++ data-remote/src/main/kotlin/com/bff/wespot/data/remote/model/auth/request/SignUpDto.kt
@@ -1,0 +1,14 @@
+package com.bff.wespot.data.remote.model.auth.request
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class SignUpDto(
+    val name: String,
+    val schoolId: Int,
+    val grade: Int,
+    val classNumber: Int,
+    val gender: String,
+    val signUpToken: String,
+    val consents: ConsentsDto,
+)
