+++ data-remote/src/main/kotlin/com/bff/wespot/data/remote/model/auth/response/AuthTokenDto.kt
@@ -1,0 +1,17 @@
+package com.bff.wespot.data.remote.model.auth.response
+
+import com.bff.wespot.model.auth.response.AuthToken
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class AuthTokenDto(
+    val accessToken: String,
+    val refreshToken: String,
+    val refreshTokenExpiredAt: String,
+) {
+    fun toAuthToken() = AuthToken(
+        accessToken = accessToken,
+        refreshToken = refreshToken,
+        refreshTokenExpiredAt = refreshTokenExpiredAt,
+    )
+}
