+++ data-remote/src/main/kotlin/com/bff/wespot/data/remote/model/auth/response/SchoolListDto.kt
@@ -1,0 +1,8 @@
+package com.bff.wespot.data.remote.model.auth.response
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class SchoolListDto(
+    val schools: List<SchoolDto>
+)
