+++ data-remote/src/main/kotlin/com/bff/wespot/data/remote/model/message/request/MessageTypeDto.kt
@@ -1,0 +1,8 @@
+package com.bff.wespot.data.remote.model.message.request
+
+enum class MessageTypeDto {
+    RECEIVED,
+    SENT,
+}
+
+internal fun MessageTypeDto.type() = this.name.lowercase()
