+++ data-remote/src/main/kotlin/com/bff/wespot/data/remote/model/message/request/SentMessageDto.kt
@@ -1,0 +1,13 @@
+package com.bff.wespot.data.remote.model.message.request
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class SentMessageDto (
+    val receivedId: Int,
+    val schoolId: Int,
+    val grade: Int,
+    val group: Int,
+    val content: String,
+    val sender: String,
+)
