+++ data-remote/src/main/kotlin/com/bff/wespot/data/remote/model/message/response/MessageIdDto.kt
@@ -1,0 +1,8 @@
+package com.bff.wespot.data.remote.model.message.response
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class MessageIdDto (
+    val id: String,
+)
