+++ data-remote/src/main/kotlin/com/bff/wespot/data/remote/source/message/MessageDataSource.kt
@@ -1,0 +1,15 @@
+package com.bff.wespot.data.remote.source.message
+
+import com.bff.wespot.data.remote.model.message.request.MessageTypeDto
+import com.bff.wespot.data.remote.model.message.request.SentMessageDto
+import com.bff.wespot.data.remote.model.message.response.MessageIdDto
+import com.bff.wespot.data.remote.model.message.response.MessageListDto
+import com.bff.wespot.data.remote.model.message.response.MessageStatusDto
+
+interface MessageDataSource {
+    suspend fun getMessageList(messageTypeDto: MessageTypeDto): Result<MessageListDto>
+
+    suspend fun postMessage(sentMessageDto: SentMessageDto): Result<MessageIdDto>
+
+    suspend fun getMessageStatus(): Result<MessageStatusDto>
+}
