+++ data/src/main/kotlin/com/bff/wespot/data/mapper/auth/AuthMapper.kt
@@ -1,14 +1,13 @@
 package com.bff.wespot.data.mapper.auth
 
+import com.bff.wespot.data.remote.model.auth.request.ConsentsDto
+import com.bff.wespot.data.remote.model.auth.request.SignUpDto
 import com.bff.wespot.model.auth.request.KakaoAuthToken
 import com.bff.wespot.model.auth.request.SignUp
 import com.bff.wespot.model.auth.response.Consents
-import com.bff.wespot.network.model.auth.request.ConsentsDto
-import com.bff.wespot.network.model.auth.request.KakaoAuthTokenDto
-import com.bff.wespot.network.model.auth.request.SignUpDto
 
 internal fun KakaoAuthToken.toDto() =
-    KakaoAuthTokenDto(
+    com.bff.wespot.data.remote.model.auth.request.KakaoAuthTokenDto(
         socialType = socialType,
         identityToken = idToken ?: "",
     )
