+++ data/src/main/kotlin/com/bff/wespot/data/mapper/message/MessageMapper.kt
@@ -2,8 +2,8 @@ package com.bff.wespot.data.mapper.message
 
 import com.bff.wespot.model.message.request.MessageType
 import com.bff.wespot.model.message.request.SentMessage
-import com.bff.wespot.network.model.message.request.MessageTypeDto
-import com.bff.wespot.network.model.message.request.SentMessageDto
+import com.bff.wespot.data.remote.model.message.request.MessageTypeDto
+import com.bff.wespot.data.remote.model.message.request.SentMessageDto
 
 internal fun MessageType.toMessageTypeDto(): MessageTypeDto {
     return when(this) {
