+++ settings.gradle.kts
@@ -29,6 +29,7 @@ include(":feature:auth")
 include(":feature:message")
 include(":domain")
 include(":data")
+include(":data-remote")
 include(":designsystem")
 include(":core:model")
 include(":core:common")
